+++ spec/views/catalog/_resource_attributes_default.html.erb_spec.rb
@@ -171,6 +171,31 @@ RSpec.describe "catalog/_resource_attributes_default.html.erb" do
       expect(rendered).to have_selector "span.label-primary.fixity-count", text: "1"
     end
   end
+
+  context "when given a ScannedResource with a component id" do
+    let(:scanned_resource) do
+      FactoryBot.create_for_repository(
+        :scanned_resource,
+        source_metadata_identifier: "AC044_c0003"
+      )
+    end
+    let(:document) { Valkyrie::MetadataAdapter.find(:index_solr).resource_factory.from_resource(resource: scanned_resource) }
+    let(:solr_document) { SolrDocument.new(document) }
+
+    before do
+      stub_pulfa(pulfa_id: "AC044_c0003")
+      assign :document, solr_document
+      allow(view).to receive(:document).and_return(solr_document)
+      allow(view).to receive(:has_search_parameters?).and_return(false)
+      stub_blacklight_views
+      render
+    end
+    it "provides a link to the finding aid" do
+      expect(rendered).to have_selector "th", text: "Source Metadata Identifier"
+      expect(rendered).to have_link "AC044_c0003", href: "https://findingaids.princeton.edu/collections/AC044/c0003"
+    end
+  end
+
   context "when given a ScannedMap solr document" do
     let(:scanned_map) do
       FactoryBot.create_for_repository(:scanned_map,